user model error improvement

+++ src/backend/model/user/model.ts
@@ -14,11 +14,11 @@ const model = {
   user: {
     create: async function (data: IUser): Promise<IRefreshToken> {
       try {
-        if (data._id !== undefined) throw new UserBadRequest('You cant put the _id yourself')
-        if (data.refreshToken !== undefined) throw new UserBadRequest('You cant put the refreshToken yourself')
+        if (data._id !== undefined) throw new UserBadRequest('Invalid credentials', 'You can not put the _id yourself')
+        if (data.refreshToken !== undefined) throw new UserBadRequest('Invalid credentials', 'You can not put the refreshToken yourself')
 
         const isValidAccount = verifyEmail(data.account)
-        if (!isValidAccount) throw new UserBadRequest('Invalid account it must match example@service.ext')
+        if (!isValidAccount) throw new UserBadRequest('Invalid credentials', 'Invalid account it must match example@service.ext')
 
         const salt = await bcrypt.genSalt(Number(BCRYPT_SALT_HASH))
         const hashedPwd = await bcrypt.hash(data.pwd, salt)
@@ -28,15 +28,15 @@ const model = {
           { pwd: 0, refreshToken: 0 }
         ).lean()
 
-        if (user === null) throw new NotFound('User do not exist')
+        if (user === null) throw new NotFound('User not found')
 
         return user
       } catch (e: any) {
-        if (e?.code === 11000) throw new DuplicateData('This user already exists')
+        if (e?.code === 11000) throw new DuplicateData('User already exists')
         else if (e instanceof UserBadRequest) throw e
         else if (e instanceof NotFound) throw e
 
-        throw new DatabaseError('Something went wrong while writing the user')
+        throw new DatabaseError('Failed to save', 'The user was not created, something went wrong please try again')
       }
     },
     update: async function (data: Partial<IUser>, userId: Types.ObjectId): Promise<IRefreshToken> {
@@ -46,45 +46,49 @@ const model = {
         data.pwd = pwd
       }
 
-      if (data.account !== undefined) throw new UserBadRequest('You need to use the endpoint for account change')
-      if (data._id !== undefined) throw new UserBadRequest('You cant change the _id field')
-      if (data.refreshToken !== undefined) throw new UserBadRequest('You cant change the refreshToken field')
+      if (data.account !== undefined) throw new UserBadRequest('Invalid credentials', 'You can not update the account here')
+      if (data._id !== undefined) throw new UserBadRequest('Invalid credentials', 'You can not change the _id')
+      if (data.refreshToken !== undefined) throw new UserBadRequest('Invalid credentials', 'You can not update the refreshToken')
 
       const user = await dbModel.updateOne({ _id: userId }, { ...data })
-      if (user.matchedCount === 0) throw new NotFound('User does not exist')
+      if (user.matchedCount === 0) throw new NotFound('User not found')
       const refreshToken = await dbModel.findOne({ _id: userId }, { pwd: 0, refreshToken: 0 }).lean()
 
       if (user.acknowledged && refreshToken !== null) {
         return refreshToken
       }
 
-      throw new NotFound('User does not exist')
+      throw new NotFound('User not found')
     },
     delete: async function (userId: Types.ObjectId): Promise<boolean> {
+      if (!Types.ObjectId.isValid(userId)) {
+        throw new UserBadRequest('Invalid credentials', 'The _id is invalid')
+      }
+
       const result = await dbModel.deleteOne({ _id: userId })
 
       if (result.acknowledged && result.deletedCount === 1) { return true }
-      throw new NotFound('User may not exist or the id is incorrect')
+      throw new NotFound('User not found')
     },
     account: {
       update: async function (userId: Types.ObjectId, account: string): Promise<IRefreshToken> {
         const isValidAccount = verifyEmail(account)
-        if (!isValidAccount) throw new UserBadRequest('Invalid account it must match example@service.ext')
+        if (!isValidAccount) throw new UserBadRequest('Invalid credentials', 'The account must match example@service.ext')
 
         const response = await dbModel.updateOne({ _id: userId }, { account })
-        if (response.matchedCount === 0) throw new NotFound('User does not exist')
+        if (response.matchedCount === 0) throw new NotFound('User not found')
 
         const user = await dbModel.findOne({ _id: userId }, { refreshToken: 0, pwd: 0 }).lean()
-        if (user === null) throw new NotFound('User does not exist')
+        if (user === null) throw new NotFound('User not found')
         return user
       }
     },
     password: {
       update: async function (account: string, pwd: string): Promise<boolean> {
-        if (!verifyEmail(account)) throw new UserBadRequest('invalid account')
+        if (!verifyEmail(account)) throw new UserBadRequest('Invalid credentials', 'The account must match example@service.ext')
 
         const response = await dbModel.updateOne({ account }, { pwd })
-        if (response.matchedCount === 0) throw new NotFound('User does not exist')
+        if (response.matchedCount === 0) throw new NotFound('User not found')
 
         return true
       }
